% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/read-data.R
\name{create_position_set}
\alias{create_position_set}
\title{create_position_set() function}
\usage{
create_position_set(targets = data.frame(), regions = data.frame(),
  polymorphisms = data.frame())
}
\arguments{
\item{regions}{bed file}
}
\description{
This function uses targets, regions and polymorphism info and creates a PositionSet instance
}
\examples{
create_position_set(targets,regions,polymorphisms)
}
\keyword{PositionSet}
\keyword{instance}

