% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/methods-PositionSet.R
\name{get_variation}
\alias{get_variation}
\title{get_variation function}
\usage{
get_variation(object)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
get_variation()
}
\keyword{cats}

