% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/methods-PositionSet.R
\name{graph_barerror}
\alias{graph_barerror}
\title{Graph_barerror function}
\usage{
graph_barerror(object, depth, samples = object@samples$sample, zoom, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
graph_barerror()
}
\keyword{cats}

