% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/methods-PositionSet.R
\name{mean_sd}
\alias{mean_sd}
\title{mean_sd function}
\usage{
mean_sd(object, position, samples, nucleotide)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
mean_sd()
}
\keyword{cats}

