% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_regions}
\alias{read_regions}
\title{read_regions() function}
\usage{
read_regions(file = "regions.bed")
}
\arguments{
\item{regions}{file}
}
\description{
This function reads regions file
}
\examples{
read_targets(file="regions.bed")
}
\keyword{regions}

