% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_targets}
\alias{read_targets}
\title{read_targets() function}
\usage{
read_targets(file = "targets.txt")
}
\arguments{
\item{targets}{file}
}
\description{
This function reads targets file
}
\examples{
read_targets(file="targets.txt")
}
\keyword{targets}

